+++ src/libgpaste/keybinder/gpaste-keybinding.h
@@ -41,7 +41,7 @@ typedef const gchar *(*GPasteKeybindingGetter) (const GPasteSettings *settings);
 typedef void         (*GPasteKeybindingFunc)   (GPasteKeybinding *self,
                                                 gpointer          data);
 
-const guint    *g_paste_keybinding_get_keycodes    (const GPasteKeybinding *self);
+const guint32  *g_paste_keybinding_get_keycodes    (const GPasteKeybinding *self);
 GdkModifierType g_paste_keybinding_get_modifiers   (const GPasteKeybinding *self);
 const gchar    *g_paste_keybinding_get_dconf_key   (const GPasteKeybinding *self);
 const gchar    *g_paste_keybinding_get_accelerator (const GPasteKeybinding *self,
@@ -54,7 +54,7 @@ gboolean        g_paste_keybinding_is_active     (GPasteKeybinding *self);
 void            g_paste_keybinding_perform       (GPasteKeybinding *self);
 void            g_paste_keybinding_notify        (GPasteKeybinding *self,
                                                   GdkModifierType   modifiers,
-                                                  guint             keycode);
+                                                  guint64           keycode);
 
 GPasteKeybinding *g_paste_keybinding_new (GType                  type,
                                           const gchar           *dconf_key,
