+++ src/libgpaste/screensaver-client/gpaste-screensaver-client.c
@@ -49,7 +49,7 @@ enum
     LAST_SIGNAL
 };
 
-static guint signals[LAST_SIGNAL] = { 0 };
+static guint64 signals[LAST_SIGNAL] = { 0 };
 
 static void
 g_paste_screensaver_client_g_signal (GDBusProxy  *proxy,
