+++ src/libgpaste/ui/gpaste-ui-item-action.h
@@ -39,11 +39,11 @@ struct _GPasteUiItemActionClass
     /*< pure virtual >*/
     void (*activate) (GPasteUiItemAction *self,
                       GPasteClient       *client,
-                      guint32             index);
+                      guint64             index);
 };
 
 void g_paste_ui_item_action_set_index (GPasteUiItemAction *self,
-                                       guint32             index);
+                                       guint64             index);
 
 GtkWidget *g_paste_ui_item_action_new (GType         type,
                                        GPasteClient *client,
