+++ src/libgpaste/ui/gpaste-ui-item.h
@@ -35,12 +35,12 @@ G_PASTE_FINAL_TYPE (UiItem, ui_item, UI_ITEM, GPasteUiItemSkeleton)
 void g_paste_ui_item_activate  (GPasteUiItem *self);
 void g_paste_ui_item_refresh   (GPasteUiItem *self);
 void g_paste_ui_item_set_index (GPasteUiItem *self,
-                                guint32       index);
+                                guint64       index);
 
 GtkWidget *g_paste_ui_item_new (GPasteClient   *client,
                                 GPasteSettings *settings,
                                 GtkWindow      *rootwin,
-                                guint32         index);
+                                guint64         index);
 
 G_END_DECLS
